refactor(lib): split three 1k+ modules into domain files (77→75)
Chips away at the FILES_OVER_1K_BUDGET backlog with behavior-neutral
extractions; every consumer keeps its import path via re-exports.

- stickToBottom.test.ts (1060) → split scroll-release / escape /
  pin-state / programmatic-leave half into stickToBottom.core.test.ts.
- contextUsage.ts (1381) → per-session localStorage usage snapshot
  moves to contextUsageSnapshot.ts (save / load / clear /
  restoreContextUsageForSession).
- goalOrch.ts (1139) → session indicator + view assembly into
  goalOrchView.ts; UI display pref (parse / load / save + change
  event) into goalOrchUiPref.ts.

No product behavior change. pnpm typecheck, lint, test (6870),
build:ui, cargo test (1580), and check-code-quality-gates --mode
final all pass; files ≥1k lines 77 → 75.
